+++ NEWS.md
@@ -2,4 +2,4 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 
-* Added one new feature, namely, it saves the RLH values for each unit to the choicemodelr output object and to an RLH.csv file (saved to current working directory).
+* two new features to version 1.2: (1) Files RBetas.csv, RLog.txt, and restart.txt are saved to a directory specified by the argument, directory, of the choicemodelr function; (2) choicemodelr also saves the RLH values for each unit to the choicemodelr output object and to an RLH.csv file (saved to the specified directory).